Debugged metal sinogram estimation
